# [internal/stanza] Batch in emitter instead of converter