--- a/internal/stanza/converter.go
+++ b/internal/stanza/converter.go
@@ -24,31 +24,21 @@ import (
 	"runtime"
 	"sort"
 	"sync"
-	"time"
 
 	"github.com/open-telemetry/opentelemetry-log-collection/entry"
 	"go.opentelemetry.io/collector/model/pdata"
 	"go.uber.org/zap"
 )
 
-const (
-	// DefaultFlushInterval is the default flush interval.
-	DefaultFlushInterval = 100 * time.Millisecond
-	// DefaultMaxFlushCount is the default max flush count.
-	DefaultMaxFlushCount = 100
-)
-
-// Converter converts entry.Entry into pdata.Logs aggregating translated
+// Converter converts a batch of entry.Entry into pdata.Logs aggregating translated
 // entries into logs coming from the same Resource.
-// Logs are being sent out based on the flush interval and/or the maximum
-// batch size.
 //
 // The diagram below illustrates the internal communication inside the Converter:
 //
 //            ┌─────────────────────────────────┐
 //            │ Batch()                         │
-//  ┌─────────┤  Ingests log entries and sends  │
-//  │         │  them onto a workerChan         │
+//  ┌─────────┤  Ingests batches of log entries │
+//  │         │  and sends them onto workerChan │
 //  │         └─────────────────────────────────┘
 //  │
 //  │ ┌───────────────────────────────────────────────────┐
@@ -59,16 +49,15 @@ const (
 //  └─┼─┼─► workerLoop()                                      │
 //    └─┤ │   consumes sent log entries from workerChan,      │
 //      │ │   translates received entries to pdata.LogRecords,│
-//      └─┤   marshalls them to JSON and send them onto       │
-//        │   batchChan                                       │
+//      └─┤   hashes them to generate an ID, and sends them   │
+//        │   onto batchChan                                  │
 //        └─────────────────────────┬─────────────────────────┘
 //                                  │
 //                                  ▼
 //      ┌─────────────────────────────────────────────────────┐
-//      │ batchLoop()                                         │
+//      │ aggregationLoop()                                   │
 //      │   consumes from batchChan, aggregates log records   │
-//      │   by marshaled Resource and based on flush interval │
-//      │   and maxFlushCount decides whether to send the     │
+//      │   by marshaled Resource and sends the               │
 //      │   aggregated buffer to flushChan                    │
 //      └───────────────────────────┬─────────────────────────┘
 //                                  │
@@ -81,7 +70,7 @@ const (
 //      └─────────────────────────────────────────────────────┘
 //
 type Converter struct {
-	// pLogsChan is a channel on which batched logs will be sent to.
+	// pLogsChan is a channel on which aggregated logs will be sent to.
 	pLogsChan chan pdata.Logs
 
 	stopOnce sync.Once
@@ -92,11 +81,10 @@ type Converter struct {
 	workerChan chan []*entry.Entry
 	// workerCount configures the amount of workers started.
 	workerCount int
-	// batchChan obtains log entries converted by the pool of workers,
-	// in a form of logRecords grouped by Resource and then after aggregating
-	// them decides based on maxFlushCount if the flush should be triggered.
-	// If also serves the ticker flushes configured by flushInterval.
-	batchChan chan []workerItem
+	// aggregationChan obtains log entries converted by the pool of workers,
+	// in a form of logRecords grouped by Resource and then sends aggregated logs
+	// on flushChan.
+	aggregationChan chan []workerItem
 
 	// flushChan is an internal channel used for transporting batched pdata.Logs.
 	flushChan chan pdata.Logs
@@ -135,14 +123,14 @@ func WithWorkerCount(workerCount int) ConverterOption {
 
 func NewConverter(opts ...ConverterOption) *Converter {
 	c := &Converter{
-		workerChan:  make(chan []*entry.Entry),
-		workerCount: int(math.Max(1, float64(runtime.NumCPU()/4))),
-		batchChan:   make(chan []workerItem),
-		data:        make(map[uint64]pdata.Logs),
-		pLogsChan:   make(chan pdata.Logs),
-		stopChan:    make(chan struct{}),
-		logger:      zap.NewNop(),
-		flushChan:   make(chan pdata.Logs),
+		workerChan:      make(chan []*entry.Entry),
+		workerCount:     int(math.Max(1, float64(runtime.NumCPU()/4))),
+		aggregationChan: make(chan []workerItem),
+		data:            make(map[uint64]pdata.Logs),
+		pLogsChan:       make(chan pdata.Logs),
+		stopChan:        make(chan struct{}),
+		logger:          zap.NewNop(),
+		flushChan:       make(chan pdata.Logs),
 	}
 
 	for _, opt := range opts {
@@ -161,7 +149,7 @@ func (c *Converter) Start() {
 	}
 
 	c.wg.Add(1)
-	go c.batchLoop()
+	go c.aggregationLoop()
 
 	c.wg.Add(1)
 	go c.flushLoop()
@@ -188,7 +176,7 @@ type workerItem struct {
 
 // workerLoop is responsible for obtaining log entries from Batch() calls,
 // converting them to pdata.LogRecords and sending them together with the
-// associated Resource through the batchChan for aggregation.
+// associated Resource through the aggregationChan for aggregation.
 func (c *Converter) workerLoop() {
 	defer c.wg.Done()
 
@@ -216,22 +204,21 @@ func (c *Converter) workerLoop() {
 			}
 
 			select {
-			case c.batchChan <- workerItems:
+			case c.aggregationChan <- workerItems:
 			case <-c.stopChan:
 			}
 		}
 	}
 }
 
-// batchLoop is responsible for receiving the converted log entries and aggregating
+// aggregationLoop is responsible for receiving the converted log entries and aggregating
 // them by Resource.
-// Whenever maxFlushCount is reached or the ticker ticks a flush is triggered.
-func (c *Converter) batchLoop() {
+func (c *Converter) aggregationLoop() {
 	defer c.wg.Done()
 
 	for {
 		select {
-		case workerItems, ok := <-c.batchChan:
+		case workerItems, ok := <-c.aggregationChan:
 			if !ok {
 				return
 			}