--- a/internal/stanza/emitter.go
+++ b/internal/stanza/emitter.go
@@ -17,8 +17,10 @@ package stanza // import "github.com/open-telemetry/opentelemetry-collector-cont
 import (
 	"context"
 	"sync"
+	"time"
 
 	"github.com/open-telemetry/opentelemetry-log-collection/entry"
+	"github.com/open-telemetry/opentelemetry-log-collection/operator"
 	"github.com/open-telemetry/opentelemetry-log-collection/operator/helper"
 	"go.uber.org/zap"
 )
@@ -26,31 +28,134 @@ import (
 // LogEmitter is a stanza operator that emits log entries to a channel
 type LogEmitter struct {
 	helper.OutputOperator
-	logChan  chan *entry.Entry
-	stopOnce sync.Once
+	logChan       chan []*entry.Entry
+	stopOnce      sync.Once
+	cancel        context.CancelFunc
+	batchMux      sync.Mutex
+	batch         []*entry.Entry
+	wg            sync.WaitGroup
+	maxBatchSize  uint
+	flushInterval time.Duration
+}
+
+type LogEmitterOption interface {
+	Apply(*LogEmitter)
+}
+
+type logEmitterOptionFunc func(*LogEmitter)
+
+func (leo logEmitterOptionFunc) Apply(le *LogEmitter) {
+	leo(le)
+}
+
+var (
+	defaultFlushInterval      = 100 * time.Millisecond
+	defaultMaxBatchSize  uint = 200
+)
+
+// LogEmitterWithMaxBatchSize returns an option that makes the LogEmitter use the specified max batch size
+func LogEmitterWithMaxBatchSize(maxBatchSize uint) LogEmitterOption {
+	return logEmitterOptionFunc(func(le *LogEmitter) {
+		le.maxBatchSize = maxBatchSize
+		le.batch = make([]*entry.Entry, 0, maxBatchSize)
+	})
+}
+
+// LogEmitterWithFlushInterval returns an option that makes the LogEmitter use the specified flush interval
+func LogEmitterWithFlushInterval(flushInterval time.Duration) LogEmitterOption {
+	return logEmitterOptionFunc(func(le *LogEmitter) {
+		le.flushInterval = flushInterval
+	})
+}
+
+// LogEmitterWithLogger returns an option that makes the LogEmitter use the specified logger
+func LogEmitterWithLogger(logger *zap.SugaredLogger) LogEmitterOption {
+	return logEmitterOptionFunc(func(le *LogEmitter) {
+		le.OutputOperator.BasicOperator.SugaredLogger = logger
+	})
 }
 
 // NewLogEmitter creates a new receiver output
-func NewLogEmitter(logger *zap.SugaredLogger) *LogEmitter {
-	return &LogEmitter{
+func NewLogEmitter(opts ...LogEmitterOption) *LogEmitter {
+	le := &LogEmitter{
 		OutputOperator: helper.OutputOperator{
 			BasicOperator: helper.BasicOperator{
 				OperatorID:    "log_emitter",
 				OperatorType:  "log_emitter",
-				SugaredLogger: logger,
+				SugaredLogger: zap.NewNop().Sugar(),
 			},
 		},
-		logChan: make(chan *entry.Entry),
+		logChan:       make(chan []*entry.Entry),
+		maxBatchSize:  defaultMaxBatchSize,
+		batch:         make([]*entry.Entry, 0, defaultMaxBatchSize),
+		flushInterval: defaultFlushInterval,
 	}
+
+	for _, opt := range opts {
+		opt.Apply(le)
+	}
+
+	return le
+}
+
+func (e *LogEmitter) Start(_ operator.Persister) error {
+	ctx, cancel := context.WithCancel(context.Background())
+	e.cancel = cancel
+
+	e.wg.Add(1)
+	go e.flusher(ctx)
+	return nil
 }
 
 // Process will emit an entry to the output channel
 func (e *LogEmitter) Process(ctx context.Context, ent *entry.Entry) error {
+	e.batchMux.Lock()
+
+	e.batch = append(e.batch, ent)
+	if uint(len(e.batch)) >= e.maxBatchSize {
+		// flushTriggerAmount triggers a blocking flush
+		e.batchMux.Unlock()
+		e.flush(ctx)
+		return nil
+	}
+
+	e.batchMux.Unlock()
+	return nil
+}
+
+func (e *LogEmitter) flusher(ctx context.Context) {
+	defer e.wg.Done()
+
+	ticker := time.NewTicker(e.flushInterval)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ticker.C:
+			e.flush(ctx)
+		case <-ctx.Done():
+			return
+		}
+	}
+}
+
+func (e *LogEmitter) flush(ctx context.Context) {
+	var batch []*entry.Entry
+	e.batchMux.Lock()
+
+	if len(e.batch) == 0 {
+		e.batchMux.Unlock()
+		return
+	}
+	batch = e.batch
+	e.batch = make([]*entry.Entry, 0, e.maxBatchSize)
+
+	e.batchMux.Unlock()
+
 	select {
-	case e.logChan <- ent:
+	case e.logChan <- batch:
 	case <-ctx.Done():
 	}
-	return nil
 }
 
 // Stop will close the log channel
@@ -58,5 +163,12 @@ func (e *LogEmitter) Stop() error {
 	e.stopOnce.Do(func() {
 		close(e.logChan)
 	})
+
+	if e.cancel != nil {
+		e.cancel()
+		e.cancel = nil
+	}
+
+	e.wg.Wait()
 	return nil
 }