… internal/stanza/converter_test.go
… internal/stanza/emitter_test.go
… internal/stanza/factory.go
… internal/stanza/integration_test.go
… internal/stanza/receiver_test.go
… receiver/filelogreceiver/filelog.go
… receiver/filelogreceiver/filelog_test.go