**Description:** 
* Moved batching in the internal/stanza package from the Converter to the LogEmitter to increase performance.

When the worker pattern was implemented, batching was put after the worker loop in the converter. The problem with this is that the workers have very little work to do per cycle. Ideally, workers have large amounts of work they do before they request more. Otherwise, workers begin to incur a lot of overhead from requesting work constantly (in this case selecting from a channel), which can (and does) cause performance issues under high load.

Instead of a unit of work being a single entry, this PR makes a unit of work a batch of entries, moving the batching logic into the LogEmitter.

**Testing:**
Added tests to LogEmitter to validate MaxBatchSize and FlushInterval are still respected.

Benchmarks of emitter-to-consumer performance:

Before:
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/stanza
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkEmitterToConsumer/worker_count=1-12                  10        5860113666 ns/op        3075350700 B/op 61522383 allocs/op
BenchmarkEmitterToConsumer/worker_count=2-12                  10        4330471031 ns/op        3075330906 B/op 61521719 allocs/op
BenchmarkEmitterToConsumer/worker_count=4-12                  10        3720432579 ns/op        3075325519 B/op 61521505 allocs/op
BenchmarkEmitterToConsumer/worker_count=6-12                  10        3540257550 ns/op        3075323456 B/op 61521449 allocs/op
BenchmarkEmitterToConsumer/worker_count=8-12                  10        3540537514 ns/op        3075327725 B/op 61521481 allocs/op
PASS
```

After:
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/stanza
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkEmitterToConsumer/worker_count=1-12                  10        3025573325 ns/op        3087150870 B/op 60540716 allocs/op
BenchmarkEmitterToConsumer/worker_count=2-12                  10        1677928788 ns/op        3087136036 B/op 60540400 allocs/op
BenchmarkEmitterToConsumer/worker_count=4-12                  10        1235078416 ns/op        3087130332 B/op 60540275 allocs/op
BenchmarkEmitterToConsumer/worker_count=6-12                  10        1223689228 ns/op        3087129233 B/op 60540246 allocs/op
BenchmarkEmitterToConsumer/worker_count=8-12                  10        1209764305 ns/op        3087128364 B/op 60540240 allocs/op
```