@BinaryFissionGames 2021-12-01
@djaglowski 
Here's the benchmarks with what is currently in this PR:
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/stanza
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkEmitterToConsumer/worker_count=1-12                  10        3001780091 ns/op        3087148339 B/op 60540706 allocs/op
BenchmarkEmitterToConsumer/worker_count=2-12                  10        1722811735 ns/op        3087141495 B/op 60540454 allocs/op
BenchmarkEmitterToConsumer/worker_count=4-12                  10        1263703437 ns/op        3087132067 B/op 60540290 allocs/op
BenchmarkEmitterToConsumer/worker_count=6-12                  10        1233186755 ns/op        3087133180 B/op 60540251 allocs/op
BenchmarkEmitterToConsumer/worker_count=8-12                  10        1247131258 ns/op        3087130500 B/op 60540259 allocs/op
```